+++ models/src/main/java/ru/bardinpetr/itmo/lab5/models/data/collection/WorkerCollection.java
@@ -4,5 +4,5 @@ import ru.bardinpetr.itmo.lab5.models.data.Worker;
 
 import java.util.TreeSet;
 
-public class WorkerCollection extends TreeSet<Worker> implements ISetCollection<Long, Worker> {
+public class WorkerCollection extends TreeSet<Worker> implements ISetCollection<Integer, Worker> {
 }
